requirements.txt
Flask==3.0.0
pycryptodome==3.19.0
requests==2.31.0
pyserial==3.5
